refactor(services): extract pyarrow schema inference into schema_infer module
Move the six pure-pyarrow inference helpers out of the 1675-LOC services/tables.py
into a new neutral leaf module services/schema_infer.py. The new module depends
only on fabric_dw.models, fabric_dw.types, the stdlib, and pyarrow -- it imports
no SQL, no SqlTarget, and no module-global state.

Moved to schema_infer.py:
- infer_columns_from_parquet
- infer_columns_from_csv
- infer_columns_from_json
- _json_sample_source, _json_sample_to_arrow_schema, _json_sample_keys

Shim in tables.py: the three public names are re-imported from schema_infer at
the top of tables.py so all existing callers (MCP tool handlers, CLI commands)
continue to work without any change.

load.py: _infer_columns_from_local now imports infer_columns_from_csv and
infer_columns_from_parquet from schema_infer directly, breaking the entanglement
between the two largest service modules.

test_load.py: three patch strings retargeted from fabric_dw.services.tables.*
to fabric_dw.services.schema_infer.* so the mocks intercept the right binding.

A new tests/unit/services/test_schema_infer.py confirms that all three inference
functions are importable from their new home and produce correct results.

Part of #828
